requirements.txt
yfinance>=0.2.40
pandas>=2.0.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
streamlit>=1.35.0
plotly>=5.20.0
